% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomFracToDelta.R
\name{AFtoDelta}
\alias{AFtoDelta}
\title{Convert an atom fraction to a delta value}
\usage{
AFtoDelta(AF, Rst)
}
\arguments{
\item{AF}{Atomic fraction of sample}

\item{Rst}{Isotope ratio of standard}
}
\value{
delta value (permil)
}
\description{
Convert an atom fraction to a delta value
}
\examples{
AFtoDelta(0.1, 0.003678)
}
