% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPercentTracer.R
\name{calcPercentTracer}
\alias{calcPercentTracer}
\title{Calculate the amount of isotope tracer that needs to be added to
achieve a given level of isotopic enrichment (i.e. a given delta value).}
\usage{
calcPercentTracer(deltaTarg, Rst, AFadd)
}
\arguments{
\item{Rst}{Isotope ratio of standard}

\item{AFadd}{The atom fraction of the isotopically enriched element
that will be added as tracer}

\item{deltaTarget}{Target level of enrichment,
expressed as a delta value (permil)}
}
\value{
Amount of isotopic tracer to add, expressed as a percentage of the
elemental concentration in the system prior to adding tracer.
}
\description{
Calculate the amount of isotope tracer that needs to be added to
achieve a given level of isotopic enrichment (i.e. a given delta value).
}
\examples{
calcPercentTracer(10000, 0.003678, 0.98)
}
