% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaToAtomFrac.R
\name{deltaToAtomFrac}
\alias{deltaToAtomFrac}
\title{Convert from a delta value to an atom fraction}
\usage{
deltaToAtomFrac(delta, Rst)
}
\arguments{
\item{delta}{Delta value of sample}

\item{Rst}{Isotope ratio of standard}
}
\value{
Atom fraction
}
\description{
Convert from a delta value to an atom fraction
}
\examples{
deltaToAtomFrac(29209.66, 0.003678)
}
