% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RToAtomFrac.R
\name{RToAtomFrac}
\alias{RToAtomFrac}
\title{Convert from an isotope ratio to an atom fraction}
\usage{
RToAtomFrac(R)
}
\arguments{
\item{R}{Isotope ratio}
}
\value{
Atom fraction
}
\description{
Convert from an isotope ratio to an atom fraction
}
\examples{
RToAtomFrac(0.003678)
}
