% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RToDelta.R
\name{RToDelta}
\alias{RToDelta}
\title{Convert isotope ratio of sample to a delta value}
\usage{
RToDelta(R, Rst)
}
\arguments{
\item{R}{Isotope ratio of sample}

\item{Rst}{Isotope ratio of standard}
}
\value{
delta value (permil)
}
\description{
Convert isotope ratio of sample to a delta value
}
\examples{
RToDelta(0.11, 0.003678)
}
